+++ .eslintrc.js
@@ -2,17 +2,17 @@ module.exports = {
   root: true,
   extends: [
     'airbnb-base',
-    'plugin:jest/all'
+    'plugin:jest/all',
+    'plugin:sonarjs/recommended',
   ],
-  plugins: [],
+  plugins: [
+    'sonarjs',
+  ],
   env: {
     node: true,
   },
   rules: {
     'implicit-arrow-linebreak': 0,
-    'jest/no-hooks': 0,
-    'no-param-reassign': 0,
-    'no-underscore-dangle': 0,
     'import/no-extraneous-dependencies': [
       'error',
       {
@@ -22,6 +22,14 @@ module.exports = {
           'test/**/*.js'
         ]
       }
-    ]
+    ],
+    'jest/no-hooks': 0,
+    'no-param-reassign': 0,
+    'no-underscore-dangle': 0,
+    'sonarjs/cognitive-complexity': 1,
+    'sonarjs/no-duplicate-string': 0,
+    'sonarjs/no-duplicated-branches': 1,
+    'sonarjs/no-identical-functions': 0,
+    'sonarjs/no-same-line-conditional': 0
   },
 };